requirements.txt
discord.py>=2.2.2
numpy>=1.24.2
pytube>=12.1.3
Flask>=2.2.3